requirements.txt
Django==5.2.3
qrcode==8.2
Pillow==11.2.1
python-dotenv==1.1.0
pillow==11.2.1
google-generativeai==0.8.5
qrcode==8.2
pydid
PyJWT
faster-whisper 
ffmpeg-python
torch
